+++ workout_diary/src/services/user_service.py
@@ -6,12 +6,12 @@ import bcrypt
 class InvalidCredentialsError(Exception):
     pass
 
-
 class UserService:
 
     def __init__(self, user_repository = default_user_repository):
         self.new_user = None
         self._user_repository = user_repository
+        self.logged_in_username = None
 
     def create_user(self, new_username, new_password):
         self.new_user = self._user_repository.create(User(new_username, new_password))
@@ -29,21 +29,24 @@ class UserService:
 
         password_check = self.check_password(user, login_password)
 
-        if not user or password_check:
+        if not user or password_check == False:
             raise InvalidCredentialsError("Invalid username or password")
-
+            
         self._user = user
+        self.logged_in_username = user.username
 
         return user
 
     def hash_password(self, plain_password):
-        plain_password = str(plain_password).encode('utf-8')
-        return bcrypt.hashpw(plain_password, bcrypt.gensalt(10))
+        encoded_password = str(plain_password).encode('utf-8')
+        return bcrypt.hashpw(encoded_password, bcrypt.gensalt(10))
 
-    def check_password(self, user, plain_password):
-        plain_password = str(plain_password).encode('utf-8')
+    def check_password(self, user, plain_password):       
+        encoded_plain_password = str(plain_password).encode('utf-8')
+        return bcrypt.checkpw(encoded_plain_password, user.password)
 
-        return bcrypt.checkpw(plain_password, user.password)
+    def get_logged_in_username(self):
+        return self.logged_in_username
         
 
 user_service = UserService()