+++ workout_diary/src/services/workout_service.py
@@ -1,6 +1,7 @@
 from entities.user import User
 from entities.workout import Workout
 import datetime
+import random
 from repositories.workout_repository import (
     workout_repository as default_workout_repository)
 
@@ -14,16 +15,28 @@ class WorkoutService:
     def __str__(self) -> str:
         pass
 
-    def create_workout(self, name: str, user: User, date_and_time: datetime.date, repetition: bool, workout_type: bool, sets: str, details: str):
-        self.new_workout = self._workout_repository.create_workout(Workout(name, user, date_and_time, repetition, workout_type, sets, details))
+    def create_workout(self, username: str, workout_name: str, date_and_time: str, repetition: bool, workout_type: str, sets: str, details: str):
 
+        i = 0
+        for i in range(5):
+
+            workout_id = random.randint(1, 20000)
+            if self._check_workout_id(workout_id) == False:
+                workout_to_database = Workout(workout_id, username, workout_name, date_and_time, repetition, workout_type, sets, details)
+                default_workout_repository.create_workout_to_database(workout_to_database)
+                break
+            i += 1
+        #self.new_workout = Workout(workout_name, username, date_and_time, repetition, workout_type, sets, details)
+        
+
+
+    def get_workouts(self, username):
+
+        return default_workout_repository.select_workouts(username)
 
     def create_cardio_workout(self):
         #cardio = Workout
         pass
-
-    def add_weight_lifting_set(self, name, repetitions, weight):
-        self.workout_sets_weight_lifting[name] = (repetitions, weight)
 
     def delete_workout(self):
         pass
@@ -34,3 +47,11 @@ class WorkoutService:
     def adjust_weights(self, weight, percent):
         x = 1 + (percent / 100)
         return round(x * weight)
+
+    def _check_workout_id(self, workout_id):
+        if default_workout_repository.get_workout_id(workout_id) == None:
+            return False
+        else:
+            return True
+
+workout_service = WorkoutService()