+++ workout_diary/src/ui/new_workout.py
@@ -1,3 +1,0 @@
-class CreateNewWorkout:
-    def __init__(self) -> None:
-        pass